prepend shortened shebang only to executables
The OS should only interpret shebangs, if a file is executable. So there
should not be a need to modify files where no execute bit is set.
This avoids a problem that I ran into while trying to package our visualization
software COVISE (https://github.com/hlrs-vis/covise), which includes example
data in Tecplot format: the sbang post-install hook is applied to every
installed file that starts with the two characters #!, but this fails on the
binary Tecplot files, as they happen to start with #!TDV. Decoding them with
UTF-8 fails and an exception is thrown during post_install.